feat(PointsGenerator): create PointsGenerator adding check validation method
